feat: move the historical_exporter into a separate brownie script and run it in parallel to the regular exporter.
